Added 'v' command to display script version. This can be queried by serial drivers to determine the scripts name/version and imply capabilities. Examples would be: Allow a serial driver to determine this sketch is RF12demo.10 and therefore imply that it supports the:   band,group,node,header,data> sending format, as well as:   byte,byte,byte,byte,node,s syntax.
This is especially important for serial connections where the connection may not be able to be be reset (lack of RST etc) for example just connecting PWR/GND TX/RX of a gpio.

This appears to be a non-breaking change.

I would be happy for sketch to report .10 still, but a .11 would also be fine on my part. I would also be happy if you implement this in a different way to achieve the same result.

--lightbulb