Attempt to clarify SPI_SS
Arduino's digital-10, i.e. PB2 is the slave select. It has to be set up
as an output for the ATmega to use the SPI hardware in master mode. This
is independent of whether it's actually used for SPI or for a specific
SPI device. The confusion starts when another pin is used as slave
select on the bus - then PB2 still has to be defined as an output.